[MIG] product_attribute_priority from 8.0
