+++ __openerp__.py
@@ -3,7 +3,7 @@
 
 {
     "name": "Product attribute priority",
-    "version": "8.0.1.0.0",
+    "version": "9.0.1.0.0",
     "depends": [
         "product",
     ],
