+++ models/__init__.py
@@ -1,1 +1,3 @@
-from . import product
+from . import product_attribute
+from . import product_attribute_line
+from . import product_attribute_value
