+++ views/product_view.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-<data>
+<odoo>
 
     <record model="ir.ui.view" id="attribute_tree_view">
         <field name="model">product.attribute</field>
@@ -23,5 +22,4 @@
         </field>
     </record>
 
-</data>
-</openerp>
+</odoo>
